+++ ocean_lib/services/test/test_service.py
@@ -140,7 +140,7 @@ def test_inexistent_removals():
     ddo_dict = get_sample_ddo_with_compute_service()
     del ddo_dict["services"][1]["compute"]["publisherTrustedAlgorithms"]
     ddo = Asset.from_dict(ddo_dict)
-    compute_service = ddo.get_service("compute")
+    compute_service = ddo.services[1]
 
     with pytest.raises(
         ValueError, match="Algorithm notadid is not in trusted algorithms"
@@ -150,9 +150,66 @@ def test_inexistent_removals():
     ddo_dict = get_sample_ddo_with_compute_service()
     del ddo_dict["services"][1]["compute"]["publisherTrustedAlgorithmPublishers"]
     ddo = Asset.from_dict(ddo_dict)
-    compute_service = ddo.get_service("compute")
+    compute_service = ddo.services[1]
 
     with pytest.raises(
         ValueError, match="Publisher notadid is not in trusted algorithm publishers"
     ):
         compute_service.remove_publisher_trusted_algorithm_publisher("notadid")
+
+
+@pytest.mark.unit
+def test_utilitary_functions_for_trusted_algorithms(publisher_ocean_instance):
+    """Tests adding/removing trusted algorithms in the DDO metadata."""
+    algorithm_ddo = get_sample_algorithm_ddo(filename="ddo_algorithm2.json")
+    algorithm_ddo.did = "did:op:123"
+    algorithm_ddo_v2 = get_sample_algorithm_ddo(filename="ddo_algorithm2.json")
+    algorithm_ddo_v2.did = "did:op:1234"
+    algorithm_ddo_v3 = get_sample_algorithm_ddo(filename="ddo_algorithm2.json")
+    algorithm_ddo_v3.did = "did:op:3333"
+
+    ddo = Asset.from_dict(
+        get_sample_ddo_with_compute_service("ddo_v4_with_compute_service2.json")
+    )
+
+    publisher_trusted_algorithms = [algorithm_ddo.generate_trusted_algorithms()]
+    assert len(publisher_trusted_algorithms) == 1
+    compute_service = ddo.services[1]
+    assert compute_service.type == "compute"
+    assert (
+        compute_service.compute_values["publisherTrustedAlgorithms"]
+        == publisher_trusted_algorithms
+    )
+
+    new_publisher_trusted_algorithms = compute_service.add_publisher_trusted_algorithm(
+        algorithm_ddo_v2,
+    )
+
+    assert (
+        new_publisher_trusted_algorithms is not None
+    ), "Added a new trusted algorithm failed. The list is empty."
+    assert len(new_publisher_trusted_algorithms) == 2
+
+    new_publisher_trusted_algorithms = compute_service.add_publisher_trusted_algorithm(
+        algorithm_ddo
+    )
+    assert new_publisher_trusted_algorithms is not None
+    for trusted_algorithm in publisher_trusted_algorithms:
+        assert (
+            trusted_algorithm["did"] == algorithm_ddo.did
+        ), "Added a different algorithm besides the existing ones."
+    assert len(new_publisher_trusted_algorithms) == 2
+
+
+@pytest.mark.unit
+def test_add_trusted_algorithm_no_compute_service(publisher_ocean_instance):
+    """Tests if the DDO has or not a compute service."""
+    algorithm_ddo = get_sample_algorithm_ddo("ddo_algorithm2.json")
+    algorithm_ddo.did = "did:op:0x666"
+
+    ddo = Asset.from_dict(get_sample_ddo())
+    access_service = ddo.services[0]
+    assert access_service.type == "access"
+
+    with pytest.raises(AssertionError, match="Service is not compute type"):
+        access_service.add_publisher_trusted_algorithm(algorithm_ddo)
